Increment6 newstorage
* temp updates

* interface storage

* some update

* sync update

* tests update

---------
