+++ .github/workflows/statictest.yml
