--- cmd/server/devops-metrics-db.json
@@ -1,1 +1,0 @@
-{"GMetrics":{"HeapReleased":{"Name":"HeapReleased","Value":2940930}},"CMetrics":{"RandomValue":{"Name":"RandomValue","Value":47}}}
