+++ internal/agent/json.go
