+++ internal/agent/json_test.go
