--- internal/server/handlers/json_test.go
@@ -1,24 +1,0 @@
-package handlers
-
-import (
-	"net/http"
-	"testing"
-)
-
-func TestMetricJSONHandler(t *testing.T) {
-	type args struct {
-		w http.ResponseWriter
-		r *http.Request
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		// TODO: Add test cases.
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			MetricJSONHandler(tt.args.w, tt.args.r)
-		})
-	}
-}
