+++ internal/server/server.go
@@ -1,17 +1,37 @@
 package server
 
 import (
-	"devops-tpl/internal/server/handlers"
+	"devops-tpl/internal/server/handler"
 	"devops-tpl/internal/storage"
-	"log"
-	"os"
-	"time"
+	"devops-tpl/internal/storage/filestorage"
+	"devops-tpl/internal/storage/memstorage"
+	"net/http"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 )
 
-func NewRouter() chi.Router {
+func NewRouter(db storage.Storage) chi.Router {
+	MetricJSONHandler := func(w http.ResponseWriter, r *http.Request) {
+		handler.MetricJSON(w, r, db)
+	}
+
+	MetricSummaryJSONHandler := func(w http.ResponseWriter, r *http.Request) {
+		handler.MetricSummaryJSON(w, r, db)
+	}
+
+	MetricUpdateHandler := func(w http.ResponseWriter, r *http.Request) {
+		handler.MetricUpdate(w, r, db)
+	}
+
+	MetricGetHandler := func(w http.ResponseWriter, r *http.Request) {
+		handler.MetricGet(w, r, db)
+	}
+
+	MetricSummaryHandler := func(w http.ResponseWriter, r *http.Request) {
+		handler.MetricSummary(w, r, db)
+	}
+
 	r := chi.NewRouter()
 
 	r.Use(middleware.RequestID)
@@ -20,87 +40,55 @@ func NewRouter() chi.Router {
 	r.Use(middleware.Recoverer)
 
 	r.Route("/update", func(r chi.Router) {
-		r.Post("/", handlers.MetricJSONHandler)
+		r.Post("/", MetricJSONHandler)
 		r.Route("/{mtype}/{mname}/{mvalue}", func(r chi.Router) {
-			r.Post("/", handlers.MetricUpdateHandler)
-			r.Get("/", handlers.MetricUpdateHandler)
+			r.Post("/", MetricUpdateHandler)
+			r.Get("/", MetricUpdateHandler)
 		})
 	})
 
 	r.Route("/value", func(r chi.Router) {
-		r.Post("/", handlers.MetricSummaryJSONHandler)
+		r.Post("/", MetricSummaryJSONHandler)
 		r.Route("/{mtype}/{mname}", func(r chi.Router) {
-			r.Get("/", handlers.MetricGetHandler)
+			r.Get("/", MetricGetHandler)
 		})
 	})
 
 	r.Route("/", func(r chi.Router) {
-		r.Get("/", handlers.MetricSummaryHandler)
+		r.Get("/", MetricSummaryHandler)
 	})
 
 	return r
 }
 
-func InitFeatures() {
-	cfg := GetEnvConfig()
-	if cfg.RestoreSavedData {
-		reader, err := storage.NewReader(cfg.StoreFile)
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		checkFile, err := os.Stat(cfg.StoreFile)
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		size := checkFile.Size()
-
-		if size == 0 {
-			writer, err := storage.NewWriter(GetEnvConfig().StoreFile)
-			if err != nil {
-				log.Fatal(err)
-			}
-
-			if err := writer.WriteDatabase(); err != nil {
-				log.Fatal(err)
-			}
-		}
-
-		if storage.DB, err = reader.ReadDatabase(); err != nil {
-			log.Fatal(err)
-		}
-	}
-
-	if cfg.StoreFile != " " {
-		go StoreOnDisk(cfg)
-	}
+func RunMemory() *memstorage.MemStorage {
+	storage := memstorage.New()
+	return storage
 }
 
-func StoreOnDisk(cfg Config) {
-	if cfg.StoreInterval == 0 {
-		for range storage.MetricUpdated {
-			writer, err := storage.NewWriter(GetEnvConfig().StoreFile)
-			if err != nil {
-				log.Fatal(err)
-			}
+func RunFile(cfg Config) *filestorage.FileStorage {
+	storage := filestorage.New(cfg.StoreFile)
+	if cfg.StoreInterval != 0 {
+		// ctx, cancel := context.WithCancel(context.Background())
+		// defer cancel()
+		// storage.SaveTicker(ctx, cfg.StoreInterval)
+		go storage.SaveTicker(cfg.StoreInterval)
 
-			if err := writer.WriteDatabase(); err != nil {
-				log.Fatal(err)
-			}
-		}
 	} else {
-		ticker := time.NewTicker(cfg.StoreInterval)
-		for range ticker.C {
-			writer, err := storage.NewWriter(GetEnvConfig().StoreFile)
-			if err != nil {
-				log.Fatal(err)
-			}
+		storage.SyncMode = true
+	}
 
-			if err := writer.WriteDatabase(); err != nil {
-				log.Fatal(err)
-			}
-		}
+	if cfg.RestoreSavedData {
+		storage.Restore(cfg.StoreFile)
+	}
+	return storage
+}
+
+func InitConfig(cfg Config) storage.Storage {
+	if cfg.StoreFile != " " {
+		return RunFile(cfg)
+	} else {
+		return RunMemory()
 	}
 
 }
